usb: widen the bandwidth calculation past 16 bits
Transfer::_CalculateBandwidth() accumulated into a uint16 while working in
nanoseconds, so the value wrapped.

Low speed overflows on its own, before any length-dependent term: the two
low speed branches start at 64060 and 64107 ns, and the per-byte term adds
roughly 6300 ns for each byte of the transfer. An 8 byte low speed control
transfer therefore lands near 116000 and wraps to about 50000, reporting
less bandwidth than an empty transfer would. Full speed isochronous wraps
once the length passes roughly 70 bytes, which is every audio packet.

The result feeds fBandwidth, which the host controller drivers use to
decide whether a periodic transfer fits in a frame. A wrapped value makes
a large transfer look nearly free, so more transfers are admitted than the
bus can carry.

Make the accumulator uint32. The nanosecond total fits comfortably; only
the intermediate was too narrow. fBandwidth itself stays uint16, holding
microseconds after the division, which is adequate for the transfer sizes
these paths see.

Found by code inspection while porting Media-OS USB work, not by a
hardware failure traced back to it, so I cannot point at a device that
visibly misbehaves because of this one. The arithmetic is checkable
without hardware, which is why it is sent on its own.

+++ src/add-ons/kernel/bus_managers/usb/Transfer.cpp
@@ -258,7 +258,10 @@ Transfer::Finished(uint32 status, size_t actualLength)
 status_t
 Transfer::_CalculateBandwidth()
 {
-	uint16 bandwidthNS;
+	// Nanoseconds, so this must be 32-bit: a low speed transfer alone starts
+	// at over 64000 and every term below adds to it, and the isochronous
+	// terms scale with the transfer length.
+	uint32 bandwidthNS;
 	uint32 type = fPipe->Type();
 
 	switch (fPipe->Speed()) {
@@ -266,11 +269,11 @@ Transfer::_CalculateBandwidth()
 		{
 			// Direction doesn't matter for highspeed
 			if (type & USB_OBJECT_ISO_PIPE)
-				bandwidthNS = (uint16)((38 * 8 * 2.083)
+				bandwidthNS = (uint32)((38 * 8 * 2.083)
 					+ (2.083 * ((uint32)(3.167 * (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			else
-				bandwidthNS = (uint16)((55 * 8 * 2.083)
+				bandwidthNS = (uint32)((55 * 8 * 2.083)
 					+ (2.083 * ((uint32)(3.167 * (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			break;
@@ -279,12 +282,12 @@ Transfer::_CalculateBandwidth()
 		{
 			// Direction does matter this time for isochronous
 			if (type & USB_OBJECT_ISO_PIPE)
-				bandwidthNS = (uint16)
+				bandwidthNS = (uint32)
 					(((fPipe->Direction() == Pipe::In) ? 7268 : 6265)
 					+ (83.54 * ((uint32)(3.167 + (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			else
-				bandwidthNS = (uint16)(9107
+				bandwidthNS = (uint32)(9107
 					+ (83.54 * ((uint32)(3.167 + (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			break;
@@ -292,11 +295,11 @@ Transfer::_CalculateBandwidth()
 		case USB_SPEED_LOWSPEED:
 		{
 			if (fPipe->Direction() == Pipe::In)
-				bandwidthNS = (uint16) (64060 + (2 * USB_BW_SETUP_LOW_SPEED_PORT_DELAY)
+				bandwidthNS = (uint32) (64060 + (2 * USB_BW_SETUP_LOW_SPEED_PORT_DELAY)
 					+ (676.67 * ((uint32)(3.167 + (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			else
-				bandwidthNS = (uint16)(64107 + (2 * USB_BW_SETUP_LOW_SPEED_PORT_DELAY)
+				bandwidthNS = (uint32)(64107 + (2 * USB_BW_SETUP_LOW_SPEED_PORT_DELAY)
 					+ (667.0 * ((uint32)(3.167 + (1.1667 * 8 * fData.length))))
 					+ USB_BW_HOST_DELAY);
 			break;
